add selector and signature extension methods to functions
