+++ common/src/lib.rs
@@ -3,11 +3,13 @@
 //! Crate for common times shared between the `ethcontract` runtime crate as and
 //! the `ethcontract-derive` crate.
 
+pub mod abiext;
 pub mod bytecode;
 pub mod errors;
 pub mod hash;
 pub mod truffle;
 
+pub use crate::abiext::FunctionExt;
 pub use crate::bytecode::Bytecode;
 pub use crate::truffle::Artifact;
 pub use ethabi::{self as abi, Contract as Abi};
